Abstract Footer into Own Component
